Check if file exist before trying to restore it

+++ static/updatePanel.sh
@@ -47,7 +47,7 @@ db_connection=$(grep "^DB_CONNECTION=" "$env_file" | cut -d '=' -f 2)
 
 if [ -z "$db_connection" ]; then
   db_connection='sqlite'
-  echo "DB_CONNECTION not found in $env_file using $db_connection as default."
+  echo "DB_CONNECTION not found in $env_file, using $db_connection as default."
 else
   echo "DB_CONNECTION is set to: $db_connection"
 fi
@@ -69,10 +69,12 @@ if [ $? -ne 0 ]; then
 fi
 echo "Backed up .env file successfully."
 
-cp -a "$install_dir/storage/app/public" "$backup_dir/storage/app/"
-if [ $? -ne 0 ]; then
-  echo "Failed to backup avatars & fonts files, aborting"
-  exitInstall 1
+if [ -f "$install_dir/storage/app/public" ]; then
+  cp -a "$install_dir/storage/app/public" "$backup_dir/storage/app/"
+  if [ $? -ne 0 ]; then
+    echo "Failed to backup avatars & fonts files, aborting"
+    exitInstall 1
+  fi
 fi
 
 if [ "$db_connection" = "sqlite" ]; then
@@ -81,7 +83,7 @@ if [ "$db_connection" = "sqlite" ]; then
   if [ -z "$db_database" ]; then
     uses_default=true
     db_database='database.sqlite'
-    echo "DB_DATABASE not found in $env_file using $db_database as default."
+    echo "DB_DATABASE not found in $env_file, using $db_database as default."
   fi
 
   if [[ "$db_database" != *.sqlite ]]; then
@@ -166,14 +168,16 @@ if [ $? -ne 0 ]; then
   exitInstall 1
 fi
 
-echo "Restoring avatars & fonts"
-cp -a "$backup_dir/storage/app/public" "$install_dir/storage/app/"
-if [ $? -ne 0 ]; then
-  echo "Failed to restore avatars & fonts files, aborting"
-  exitInstall 1
+if [ -f "$backup_dir/storage/app/public" ]; then
+  echo "Restoring avatars & fonts"
+  cp -a "$backup_dir/storage/app/public" "$install_dir/storage/app/"
+  if [ $? -ne 0 ]; then
+    echo "Failed to restore avatars & fonts files, aborting"
+    exitInstall 1
+  fi
 fi
 
-if [ "$db_connection" = "sqlite" ]; then
+if [ -f "$backup_dir/$db_database.backup" ]; then
   echo "Restoring sqlite database"
   cp -a "$backup_dir/$db_database.backup" "$install_dir/database/$db_database"
   if [ $? -ne 0 ]; then
